made it work on the new API, and save the necessary things

+++ gmusic-dl.py
@@ -1,48 +1,66 @@
 #!/usr/bin/env python
- 
+
 import os
 import sys
 import eyed3
 from goldfinch import validFileName as vfn
 from gmusicapi import Mobileclient
 import getpass
+import json
 try:
   from urllib.request import urlretrieve
 except ImportError:
   from urllib import urlretrieve
 
 
-if len(sys.argv) == 1:
-  print("usage: python gmusic-dl.py <email> <album id>")
-  sys.exit()
-
-
 def normalizePath(input):
   return vfn(input, space="keep", initCap=False).decode('utf-8').rstrip(".")  
 
 
-login = sys.argv[1]
 targetDir = os.getcwd()
-albumId = sys.argv[2]
-password = getpass.getpass()
+# targetDir = "D:/Minhas músicas"
 
 eyed3.log.setLevel("ERROR")
 
+
 api = Mobileclient(debug_logging=False)
-api.login(login, password, Mobileclient.FROM_MAC_ADDRESS)
+if os.path.exists("config.json"):
+  with open('config.json', 'r') as fp:
+    ids = json.load(fp)
+  i = 0
+  while True:
+    try:
+      api.oauth_login(ids[i])
+      break
+    except:
+      print("oops, this didn't work, trying again")
+      i+=1
+      i%=len(ids)
+else:
+  oa = api.perform_oauth()
+  api.oauth_login(Mobileclient.FROM_MAC_ADDRESS,oauth_credentials=oa)
+  devices = api.get_registered_devices()
+  ids = []
+  for i in devices:
+    ids.append(i['id'][2:])
+  with open('config.json', 'w') as fp:
+      json.dump(ids, fp)
 
-album = api.get_album_info(albumId)
-dirName = normalizePath("%s - %s" % (album["artist"], album["name"]))
-dirPath = targetDir + "/" + dirName
 
-print("downloading to directory: " + dirPath)
-if not os.path.exists(dirPath):
+
+songs = api.get_all_songs()
+for song in songs:
+  dirName = normalizePath("%s - %s" % (song["artist"], song["album"]))
+  dirPath = targetDir + "/" + dirName
+  if not os.path.exists(dirPath):
+    print("downloading to directory: " + dirPath)
     os.makedirs(dirPath)
-	
-for song in album["tracks"]:
-  url = api.get_stream_url(song_id=song["storeId"], quality="hi")
   fileName = normalizePath("%s. %s - %s.mp3" % (song["trackNumber"], song["artist"], song["title"]))
   filePath = dirPath + "/" + fileName
+  if os.path.exists(filePath):
+    print(fileName+" already exists, skipping")
+    continue
+  url = api.get_stream_url(song_id=song["id"], quality="hi")
   print("downloading: " + fileName)
   urlretrieve(url, filePath)
   
@@ -51,8 +69,8 @@ for song in album["tracks"]:
     audio.tag = eyed3.id3.Tag()
     audio.tag.file_info = eyed3.id3.FileInfo(filePath)
   audio.tag.artist = song["artist"]
-  audio.tag.album = album["name"]
-  audio.tag.album_artist = album["artist"]
+  audio.tag.album = song["album"]
+  audio.tag.album_artist = song["artist"]
   audio.tag.title = song["title"]
   audio.tag.track_num = song["trackNumber"]
   audio.tag.save()
